Ahoy! Fisherman 0.5.0 =====================
Add user key bindings support.

Recall   $fisher_home/functions  are   always  before   user
functions in  $fish_function_path. This was an  early design
decision  in order  to  prevent users  from redefining  core
functions by  mistake or by  means other than  using plugins
(recommended).  In other  words, you  are free  to create  a
plugin  that modifies  a  Fisherman core  function, but  you
can't redefine  a Fisherman function privately  by saving it
to your user config fish. If  you found a bug in a Fisherman
function,  instead  of  creating  a private  patch  send  it
upstream.  If  you  created  a  function  that  overrides  a
Fisherman  core  feature,  create  a plugin.  This  way  the
community can benefit from your  code whenever you are ready
to publish it.

By  default, Fisherman  provides no  fish_user_key_bindings,
so   if   the   user   has   already   defined   their   own
fish_user_key_bindings that one will not be affected.

Now,  plugins can  define their  own key  bindings inside  a
fish_user_key_bindings.fish or key_bindings.fish at the root
of their repository or inside a functions directory. You can
put your key  bindings inside a function or not.  If you put
it inside a function, the function  name must be the same as
the file without the .fish extension.

$fisher_config/bindings.fish   When   a  plugin   with   key
bindings  is  installed  for  the first  time  or  the  only
one  with bindings  is  uninstalled,  Fisherman will  modify
~/.config/functions/fish_user_key_bindings.fish  (or  create
it for  the first  time) and  add a single  line at  the top
of  the fish_user_key_bindings  function to  source the  new
$fisher_config/bindings.fish. All  the key  bindings defined
by the enabled/installed plugins  are concatenated and saved
to this file.

This mechanism has the following advantages:

Does not slow  down shell start. Does  not require Fisherman
to  provide  his   own  fish_user_key_bindings  by  default.
Honors  any previously  existing user  key bindings.  Allows
plugin  to  define  their   own  key  bindings  and  coexist
with  the  user's key  bindings.  If  the user  updates  his
fish_user_key_bindings, re-running the  function does update
the key bindings. Mega Refactoring

The   entire   source   code   of   Fisherman   received   a
major   revision  and   refactoring.   The  validation   and
install/uninstall mechanisms were thoroughly revised and and
broken down into smaller functions easier to test as well as
several other sub parts of the system.

Rewrite fisher  search and  remove features that  are mostly
already covered by  fisher --list and remove  the ability to
generate information  about plugins  of unknown  origin. The
decision  to remove  this feature  was based  in performance
concerns  and the  result  of thinking  about the  usability
and  whether it  was really  worth the  speed tradeoff.  The
conclusion is I would rather  have better performance and if
I need  to query a  plugins origin  I can always  use fisher
--list or fisher --list=url or fisher --list=author.

Add $fisher_update_interval to determine if the index should
update  or not  when a  search  query is  taking place.  The
default value is  10 seconds. This means the  index will not
be updated  if less than  10 seconds have elapsed  since the
last action that triggered an update in the first place. See

Improve Install/Uninstall/Update status  output. If a plugin
fails  to install  decrease the  total. If  any plugins  are
skipped because  they are already  installed in the  case of
fisher install  or available in  the cache, but  disabled in
the  case of  fisher uninstall  they are  collected into  an
array and displayed in a  new section n plugin/s skipped (a,
b, c) at the bottom of the report.

Improve test coverage.

Tightly coupled  functions were making  testing increasingly
difficult.  Most of  the test  effort was  basically testing
whether git  clone or git  pull. New separation  of concerns
makes tests  run faster and the  difficult install/uninstall
algorithms has better coverage now. Other

Now  __fisher_list  can  list  plugins  from  the  cache,  a
fishfile/bundle  and plugins  that are  installed/enabled or
disabled. This  removes __fisher_file  and combines  it with
__fisher_list. This  also removes fisher -f  and replaces it
with fisher -l <file> or fisher --list=<file>.

Rename __fisher_parse_help to __fisher_complete and have the
function create  the completions automatically.  This allows
you  to complete  your  commands with  parseable usage  help
faster.  The  original  design  was fine,  but  this  change
improves auto-complete performance so it was preferred.

Use __fisher_index_update when building file with Make. This
helps prevent  an error when  using a fish version  < 2.2.0.
See #55 #50 #48.

Add  __fisher_index_update to  update the  index and  remove
previously   undocumented   fisher  update   --index.   This
function  is  designed  to bypass  GitHub's  server  network
cache  passing  an  arbitrary  query  string  to  curl  like
$fisher_index?RANDOM_NUMBER.  This means  index updates  are
immediately available now.

Add fisher --list=url option to  display local plugin url or
path.

Add fisher  --list=bare option  to display local  plugins in
the cache without the * enabled symbol.

Prepend > to  the currently enabled theme  when using fisher
--list[=cache]. Related #49.

Prepend *  to plugin  names to  indicate they  are currently
enabled when using fisher --list[=cache]. See #49.

+++ functions/fisher_uninstall.fish
@@ -1,59 +1,65 @@
 function fisher_uninstall -d "Uninstall Plugins"
-    set -l error /dev/stderr
-    set -l items
+    set -l plugins
     set -l option
+    set -l error /dev/stderr
 
     getopts $argv | while read -l 1 2
         switch "$1"
             case _
-                set items $items $2
+                set plugins $plugins $2
 
             case f force
-                set option $option force
+                set option force
 
             case q quiet
                 set error /dev/null
 
-            case help
-                set option help
-
             case h
                 printf "usage: fisher uninstall [<plugins>] [--force] [--quiet] [--help]\n\n"
 
-                printf "    -f --force  Delete copy from cache \n"
-                printf "    -q --quiet  Enable quiet mode      \n"
-                printf "     -h --help  Show usage help        \n"
+                printf "    -f --force  Delete copy from cache\n"
+                printf "    -q --quiet  Enable quiet mode\n"
+                printf "     -h --help  Show usage help\n"
                 return
 
             case \*
-                printf "fisher: Ahoy! '%s' is not a valid option\n" $1 >& 2
+                printf "fisher: '%s' is not a valid option.\n" $1 >& 2
                 fisher_uninstall -h >& 2
                 return 1
         end
     end
 
-    switch "$option"
-        case help
-            fisher help uninstall
-            return
-    end
-
     set -l time (date +%s)
     set -l count 0
     set -l index 1
-    set -l total (count $items)
+    set -l total (count $plugins)
+    set -l skipped
 
-    if set -q items[1]
-        printf "%s\n" $items
+    if set -q plugins[1]
+
+        printf "%s\n" $plugins
     else
-        __fisher_file /dev/stdin
+        __fisher_file
 
-    end | __fisher_validate | __fisher_resolve_plugin $error | while read -l path
+    end | while read -l item path
+
+        if not set item (__fisher_plugin_validate $item)
+            printf "fisher: '%s' is not a valid name, path or url.\n" $item > $error
+            continue
+        end
+
+        if not set path (__fisher_path_from_plugin $item)
+            set total (math $total - 1)
+            printf "fisher: '%s' not found\n" $item > $error
+            continue
+        end
 
         set -l name (printf "%s\n" $path | __fisher_name)
 
-        if not contains -- $name $fisher_plugins
-            if not contains -- force $option
+        if not contains -- $name (__fisher_list $fisher_file)
+            if test -z "$option"
+                set total (math $total - 1)
+                set skipped $skipped $name
                 continue
             end
         end
@@ -69,41 +75,36 @@ function fisher_uninstall -d "Uninstall Plugins"
                 set index (math $index + 1)
         end
 
-        __fisher_plugin --disable $name $path
+        if begin not __fisher_path_is_prompt $path; or test "$name" = "$fisher_prompt"; end
 
-        git -C $path ls-remote --get-url ^ /dev/null | __fisher_validate | read -l url
+            # You can use --force to remove any plugin from the cache. If prompt A is enabled
+            # you can still uninstall prompt B using --force. This will delete B's repository
+            # from $fisher_cache.
 
-        switch force
-            case $option
-                rm -rf $path
+            __fisher_plugin_disable "$name" "$path"
+        end
+
+        if test "$option" = force
+            rm -rf $path
         end
 
         set count (math $count + 1)
-
-        set -l file $fisher_config/fishfile
-
-        if not __fisher_file $file | grep -Eq "^$name\$|^$url\$"
-            continue
-        end
-
-        set -l tmp (mktemp -t fisher.XXX)
-
-        if not sed -E '/^ *'(printf "%s|%s" $name $url | sed 's|/|\\\/|g'
-        )'([ #].*)*$/d' < $file > $tmp
-            rm -f $tmp
-            printf "fisher: Could not remove '%s' from %s\n" $name $file > $error
-            return 1
-        end
-
-        mv -f $tmp $file
     end
 
     set time (math (date +%s) - $time)
 
-    if test $count = 0
+    if test ! -z "$skipped"
+        printf "%s plugin/s skipped (%s)\n" (count $skipped) (
+            printf "%s\n" $skipped | paste -sd ' ' -) > $error
+    end
+
+    if test $count -le 0
         printf "No plugins were uninstalled.\n" > $error
         return 1
     end
+
+    __fisher_complete_reset
+    __fisher_key_bindings_reset
 
     printf "Aye! %d plugin/s uninstalled in %0.fs\n" > $error $count $time
 end
